feat: todos 상태 지정 및 addTodo 함수 추가
