--- week2/Assignment2/src/assets/todos.js
@@ -1,8 +1,0 @@
-const todos = {
-    "Hyeonggeun": ['베이킹 냠냠', '방정리'],
-    "SOPT": ['웹팟 첫 세미나', '웹팟 첫 뒤풀이', '웹팟 워크샵'],
-    "세종대": ['캡스톤 발표', 'UX/UI 디자인'],
-    "운동": ['솝케팝 번개', '요가'],
-};
-
-export default todos;
